Refactor PHP class building to be DB independant - make simpler
